# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bookmark API",
    "description": "API for bookmarking web page URL",
    "version": "0.3.0.250409"
  },
  "paths": {
    "/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "ログインしてアクセストークンを取得する",
        "operationId": "login_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForLogin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "ログインユーザ情報を取得する",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForGetCurrentUser"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/bookmarks": {
      "post": {
        "tags": [
          "bookmark"
        ],
        "summary": "Add Bookmark",
        "description": "ブックマーク追加",
        "operationId": "add_bookmark_bookmarks_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestForAddBookmark"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForAddBookmark"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bookmark"
        ],
        "summary": "Get Bookmarks",
        "description": "ブックマークリスト取得",
        "operationId": "get_bookmarks_bookmarks_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForGetBookmarkList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookmarks/{hashed_id}": {
      "patch": {
        "tags": [
          "bookmark"
        ],
        "summary": "Update Bookmark",
        "description": "ブックマーク更新",
        "operationId": "update_bookmark_bookmarks__hashed_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hashed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 64,
              "maxLength": 64,
              "pattern": "[0-9a-f]+",
              "title": "Hashed Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestForUpdateBookmark"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForUpdateBookmark"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bookmark"
        ],
        "summary": "Delete Bookmark",
        "description": "ブックマーク削除",
        "operationId": "delete_bookmark_bookmarks__hashed_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hashed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 64,
              "maxLength": 64,
              "pattern": "[0-9a-f]+",
              "title": "Hashed Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForDeleteBookmark"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bookmark"
        ],
        "summary": "Get Bookmark",
        "description": "ブックマーク取得",
        "operationId": "get_bookmark_bookmarks__hashed_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hashed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 64,
              "maxLength": 64,
              "pattern": "[0-9a-f]+",
              "title": "Hashed Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForGetBookmark"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Users",
        "description": "ユーザーリスト取得",
        "operationId": "get_users_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForGetUserList"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add User",
        "description": "ユーザー追加",
        "operationId": "add_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestForAddUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForAddUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/users/{name}": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Update User",
        "description": "ユーザー更新\n - ログインユーザー自身のname,disabled,authorityは変更できない\n - 管理者以外はログインユーザー自身の情報のみ変更可能",
        "operationId": "update_user_users__name__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "pattern": "^[a-zA-Z0-9_]+$",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestForUpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForUpdateUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get User",
        "description": "ユーザー取得",
        "operationId": "get_user_users__name__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "pattern": "^[a-zA-Z0-9_]+$",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForGetUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/version": {
      "get": {
        "tags": [
          "version"
        ],
        "summary": "Get Version",
        "description": "バージョン番号取得",
        "operationId": "get_version_version_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseForGetVersion"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthorityEnum": {
        "type": "integer",
        "enum": [
          0,
          1,
          2,
          9
        ],
        "title": "AuthorityEnum",
        "description": "権限\n[権限なし: 0]\n[読み取り可: 1]\n[読み書き可: 2]\n[管理者権限: 9]"
      },
      "Body_login_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_token_post"
      },
      "BookmarkDetail": {
        "properties": {
          "hashed_id": {
            "type": "string",
            "title": "Hashed Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "memo": {
            "type": "string",
            "title": "Memo"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "hashed_id",
          "url",
          "memo",
          "created_at",
          "updated_at",
          "tags"
        ],
        "title": "BookmarkDetail"
      },
      "BookmarkDigest": {
        "properties": {
          "hashed_id": {
            "type": "string",
            "title": "Hashed Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "memo": {
            "type": "string",
            "title": "Memo"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "hashed_id",
          "url",
          "memo",
          "created_at",
          "updated_at"
        ],
        "title": "BookmarkDigest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RequestForAddBookmark": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "memo": {
            "type": "string",
            "maxLength": 400,
            "title": "Memo"
          },
          "tags": {
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "url",
          "memo",
          "tags"
        ],
        "title": "RequestForAddBookmark",
        "examples": [
          {
            "memo": "サンプル",
            "tags": [
              "work",
              "test"
            ],
            "url": "https://exsample.com"
          }
        ]
      },
      "RequestForAddUser": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_]+$",
            "title": "Name"
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "authority": {
            "$ref": "#/components/schemas/AuthorityEnum"
          }
        },
        "type": "object",
        "required": [
          "name",
          "password",
          "authority"
        ],
        "title": "RequestForAddUser",
        "examples": [
          {
            "authority": 2,
            "name": "test_user",
            "password": "password"
          }
        ]
      },
      "RequestForUpdateBookmark": {
        "properties": {
          "memo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 100,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 10,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "RequestForUpdateBookmark",
        "examples": [
          {
            "memo": "サンプル",
            "tags": [
              "private",
              "test"
            ]
          }
        ]
      },
      "RequestForUpdateUser": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "minLength": 1,
                "pattern": "^[a-zA-Z0-9_]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 8,
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "authority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorityEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "disabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled"
          }
        },
        "type": "object",
        "title": "RequestForUpdateUser",
        "examples": [
          {
            "authority": 2,
            "disabled": false,
            "name": "test_user",
            "password": "password"
          }
        ]
      },
      "ResponseForAddBookmark": {
        "properties": {
          "hashed_id": {
            "type": "string",
            "title": "Hashed Id"
          }
        },
        "type": "object",
        "required": [
          "hashed_id"
        ],
        "title": "ResponseForAddBookmark"
      },
      "ResponseForAddUser": {
        "properties": {},
        "type": "object",
        "title": "ResponseForAddUser"
      },
      "ResponseForDeleteBookmark": {
        "properties": {},
        "type": "object",
        "title": "ResponseForDeleteBookmark"
      },
      "ResponseForGetBookmark": {
        "properties": {
          "bookmark": {
            "$ref": "#/components/schemas/BookmarkDetail"
          }
        },
        "type": "object",
        "required": [
          "bookmark"
        ],
        "title": "ResponseForGetBookmark",
        "examples": [
          {
            "bookmark": {
              "created_at": "2025-01-01 12:34:56",
              "hashed_id": "123456789012345678901234567890123456789012345678901234567890abcd",
              "memo": "サンプル",
              "tags": [
                "private",
                "test"
              ],
              "updated_at": "2025-01-01 12:34:56",
              "url": "https://exsample.com"
            }
          }
        ]
      },
      "ResponseForGetBookmarkList": {
        "properties": {
          "bookmarks": {
            "items": {
              "$ref": "#/components/schemas/BookmarkDigest"
            },
            "type": "array",
            "title": "Bookmarks"
          }
        },
        "type": "object",
        "required": [
          "bookmarks"
        ],
        "title": "ResponseForGetBookmarkList",
        "examples": [
          {
            "bookmarks": [
              {
                "created_at": "2025-01-01 12:34:56",
                "hashed_id": "123456789012345678901234567890123456789012345678901234567890abcd",
                "memo": "サンプル",
                "updated_at": "2025-01-01 12:34:56",
                "url": "https://exsample.com"
              }
            ]
          }
        ]
      },
      "ResponseForGetCurrentUser": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "authority": {
            "$ref": "#/components/schemas/AuthorityEnum"
          }
        },
        "type": "object",
        "required": [
          "name",
          "authority"
        ],
        "title": "ResponseForGetCurrentUser"
      },
      "ResponseForGetUser": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDetail"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "ResponseForGetUser",
        "examples": [
          {
            "user": {
              "authority": 2,
              "disabled": false,
              "name": "test_user"
            }
          }
        ]
      },
      "ResponseForGetUserList": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserDetail"
            },
            "type": "array",
            "title": "Users"
          }
        },
        "type": "object",
        "required": [
          "users"
        ],
        "title": "ResponseForGetUserList",
        "examples": [
          {
            "users": [
              {
                "authority": 2,
                "disabled": false,
                "name": "test_user"
              }
            ]
          }
        ]
      },
      "ResponseForGetVersion": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "ResponseForGetVersion",
        "examples": [
          {
            "version": "0.3.0.250409"
          }
        ]
      },
      "ResponseForLogin": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "ResponseForLogin"
      },
      "ResponseForUpdateBookmark": {
        "properties": {
          "updated_bookmark": {
            "$ref": "#/components/schemas/BookmarkDetail"
          }
        },
        "type": "object",
        "required": [
          "updated_bookmark"
        ],
        "title": "ResponseForUpdateBookmark",
        "examples": [
          {
            "updated_bookmark": {
              "created_at": "2025-01-01 12:34:56",
              "hashed_id": "123456789012345678901234567890123456789012345678901234567890abcd",
              "memo": "サンプル",
              "tags": [
                "private",
                "test"
              ],
              "updated_at": "2025-01-01 12:34:56",
              "url": "https://exsample.com"
            }
          }
        ]
      },
      "ResponseForUpdateUser": {
        "properties": {
          "updated_user": {
            "$ref": "#/components/schemas/UserDetail"
          }
        },
        "type": "object",
        "required": [
          "updated_user"
        ],
        "title": "ResponseForUpdateUser",
        "examples": [
          {
            "updated_user": {
              "authority": 2,
              "disabled": false,
              "name": "test_user"
            }
          }
        ]
      },
      "UserDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "authority": {
            "$ref": "#/components/schemas/AuthorityEnum"
          },
          "disabled": {
            "type": "boolean",
            "title": "Disabled"
          }
        },
        "type": "object",
        "required": [
          "name",
          "authority",
          "disabled"
        ],
        "title": "UserDetail"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "auth",
      "description": "Authentication operations"
    },
    {
      "name": "bookmark",
      "description": "Bookmark operations"
    },
    {
      "name": "user",
      "description": "User operations"
    },
    {
      "name": "version",
      "description": "API version"
    }
  ]
}