# src/controllers/user.py
from fastapi import APIRouter

from ..dao.session import SessionDepend
from ..dto.user.add import RequestForAddUser, ResponseForAddUser
from ..dto.user.get import ResponseForGetUser
from ..dto.user.get_list import ResponseForGetUserList
from ..dto.user.update import RequestForUpdateUser, ResponseForUpdateUser
from ..libs.constraints import FIELD_STRING_USERNAME
from ..libs.enum import AuthorityEnum
from ..services.auth import UserDepends
from ..usecases.user import UserUsecase

router = APIRouter()


@router.post(
    "/users",
    response_model=ResponseForAddUser,
    tags=["user"],
)
def add_user(
    req: RequestForAddUser,
    session: SessionDepend,
    user: UserDepends,
) -> ResponseForAddUser:
    """
    ユーザー追加
    """
    res = UserUsecase(
        session=session,
        user=user,
        required_authority=AuthorityEnum.ADMIN,
    ).add(req)

    return ResponseForAddUser(**res)


@router.patch(
    "/users/{name}",
    response_model=ResponseForUpdateUser,
    tags=["user"],
)
def update_user(
    name: FIELD_STRING_USERNAME,
    req: RequestForUpdateUser,
    session: SessionDepend,
    user: UserDepends,
) -> ResponseForUpdateUser:
    """
    ユーザー更新
     - ログインユーザー自身のname,disabled,authorityは変更できない
     - 管理者以外はログインユーザー自身の情報のみ変更可能
    """
    res = UserUsecase(
        session=session,
        user=user,
        required_authority=AuthorityEnum.NONE,
    ).update(req, name)

    return ResponseForUpdateUser(**res)


@router.get(
    "/users/{name}",
    response_model=ResponseForGetUser,
    tags=["user"],
)
def get_user(
    name: FIELD_STRING_USERNAME,
    session: SessionDepend,
    user: UserDepends,
) -> ResponseForGetUser:
    """
    ユーザー取得
    """
    res = UserUsecase(
        session=session,
        user=user,
        required_authority=AuthorityEnum.ADMIN,
    ).get_one(name)

    return ResponseForGetUser(**res)


@router.get(
    "/users",
    response_model=ResponseForGetUserList,
    tags=["user"],
)
def get_users(
    session: SessionDepend,
    user: UserDepends,
) -> ResponseForGetUserList:
    """
    ユーザーリスト取得
    """
    res = UserUsecase(
        session=session,
        user=user,
        required_authority=AuthorityEnum.ADMIN,
    ).get_list()

    return ResponseForGetUserList(**res)